% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bedGraph.R
\name{create_bedgraph}
\alias{create_bedgraph}
\title{Converts normalized data into bedGraph files}
\usage{
create_bedgraph(
  data,
  type = c("meth", "cov"),
  all = TRUE,
  out = ".",
  group_name = NULL,
  sample_name = NULL,
  use_cpp = TRUE
)
}
\arguments{
\item{data}{Input data, can be either:
\enumerate{
\item A list of lists containing normalized samples
\item A single data frame containing one sample
\item A character string specifying path to input file
}}

\item{type}{Character string for type of value to include in bedGraph.
Should be one of "cov" or "meth" (for coverage bedGraph or methylation rate bedGraph, respectively.)}

\item{all}{Logical. Should function be performed on all files in object. Default is TRUE.
Ignored if input is a single sample or file.}

\item{out}{Directory to write output bedGraph files to. Default is current working directory.}

\item{group_name}{Name of the group (optional)}

\item{sample_name}{Name to use for output file (optional)}

\item{use_cpp}{Logical. Should function be performed on all files in object.
Default is TRUE. Will fall back to R implementation if C++ is not available.}
}
\value{
Invisibly returns list of created bedGraph file paths
}
\description{
Converts normalized data into bedGraph files
}
\details{
Converts methylation data to bedGraph format. Can handle multiple input types:
\itemize{
\item If given a normalized data object (list of lists), processes all samples by default
\item If given a single sample data frame, processes just that sample
\item If given a file path, loads and processes that file
}
}
