% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMFnorm-package.R
\docType{package}
\name{SMFnorm-package}
\alias{SMFnorm-package}
\title{SMFnorm: What the Package Does (Title Case)}
\description{
SMFnorm is a specialized R package designed for normalizing methylation data from genome-wide footprinting, such as MAPit-CpGiant or MAPit-RRBS, NOMe-seq or Fiber-seq. It implements a sophisticated quantile-bin shrinkage normalization process that addresses technical variation between replicates while preserving biological signals.

SMFnorm is a specialized R package designed for normalizing methylation data from genome-wide footprinting, such as MAPit-CpGiant or MAPit-RRBS, NOMe-seq or Fiber-seq. It implements a sophisticated quantile-bin shrinkage normalization process that addresses technical variation between replicates while preserving biological signals.
}
\keyword{internal}
